linux 5.3.8 + xorg mod upd

+++ x11-base/xorg-server-modules/xorg-server-modules-1.20.5.sh
@@ -12,9 +12,18 @@ cp -rsfT "/pkg/main/x11-base.xorg-server.mod.${PVR}/lib$LIB_SUFFIX/xorg/modules"
 # locate packages
 for pn in `curl -s http://localhost:100/apkgdb/main?action=list | grep "^x11-"`; do
 	p=/pkg/main/${pn}
+	b=`echo "$pn" | cut -d. -f1-2`
 	t=`echo "$pn" | cut -d. -f3`
 
-	if [ -d "${p}/lib$LIB_SUFFIX/xorg/modules" ]; then
+	if [ $b = "x11-base.xorg-server-modules" ]; then
+		continue
+	fi
+
+	if [ "$t" = "dev" ]; then
+		continue
+	fi
+
+	if [ -d "${p}/lib$LIB_SUFFIX/xorg/modules" -a ! -L "${p}/lib$LIB_SUFFIX/xorg/modules" ]; then
 		# copy
 		echo " * Module: $pn"
 		cp -rsfT "${p}/lib$LIB_SUFFIX/xorg/modules" "${TARGET}"
